bug fix
git-svn-id: https://tuftsvue.svn.sourceforge.net/svnroot/tuftsvue/trunk@3886 d66e5487-c4fa-4dda-8c8e-ca23240363c0

+++ VUE2/src/edu/tufts/vue/compare/VoteAggregate.java
@@ -37,7 +37,7 @@ public class VoteAggregate extends WeightAggregate {
     
     private void computeVotes() {
         int count = getCount();
-        float threshold = count/2;
+        float threshold = (float)count/2;
         for(int i=0; i<size;i++){
             for(int j=0;j<size;j++) {
                 if(c[i][j] >= threshold) {
